fix: attribution fixes and refactoring
