+++ src/components/src/index.ts
@@ -58,6 +58,7 @@ export {default as LayerAnimationControllerFactory} from './layer-animation-cont
 export {
   Attribution,
   renderBasemapAttribution,
+  dedupeBasemapAttributions,
   default as MapContainerFactory
 } from './map-container';
 export {default as MapsLayoutFactory} from './maps-layout';
@@ -104,7 +105,6 @@ export {default as LayerPanelFactory} from './side-panel/layer-panel/layer-panel
 export {default as SingleColorPalette} from './side-panel/layer-panel/single-color-palette';
 export {default as TextLabelPanelFactory} from './side-panel/layer-panel/text-label-panel';
 export {default as ScenegraphModelSelectorFactory} from './side-panel/layer-panel/scenegraph-model-selector';
-
 
 export {default as AddLayerButtonFactory} from './side-panel/layer-panel/add-layer-button';
 export * from './side-panel/layer-panel/channel-by-value-selector';
