+++ src/components/src/map-container.tsx
@@ -2,8 +2,8 @@
 // Copyright contributors to the kepler.gl project
 
 // libraries
-import React, {Component, createRef, useMemo} from 'react';
-import styled, {withTheme, useTheme} from 'styled-components';
+import React, {Component, createRef} from 'react';
+import styled, {withTheme} from 'styled-components';
 import {Map as MapboxLegacyMap, MapRef} from 'react-map-gl/mapbox-legacy';
 import {Map as MaplibreMap} from '@vis.gl/react-maplibre';
 import {PickingInfo, MapView} from '@deck.gl/core';
@@ -17,11 +17,13 @@ import {VisStateActions, MapStateActions, UIStateActions} from '@kepler.gl/actio
 // components
 import MapPopoverFactory from './map/map-popover';
 import MapControlFactory from './map/map-control';
+import {StyledMapContainer} from './common/styled-components';
 import {
-  StyledMapContainer,
-  StyledAttribution,
-  EndHorizontalFlexbox
-} from './common/styled-components';
+  Attribution,
+  AttributionLogos,
+  renderBasemapAttribution,
+  dedupeBasemapAttributions
+} from './map/attribution';
 
 import EditorFactory from './editor/editor';
 import {AnnotationOverlay} from './annotations';
@@ -52,11 +54,9 @@ import {
   mapHasOpenStreetMapAttribution,
   getBaseMapAttributions,
   getBaseMapLibrary,
-  BaseMapLibraryConfig,
   transformRequest,
   observeDimensions,
   unobserveDimensions,
-  hasMobileWidth,
   getMapLayersFromSplitMaps,
   onViewPortChange,
   getViewportFromMapState,
@@ -68,7 +68,6 @@ import {
   getLayerBlendingParameters,
   patchDeckRendererForPostProcessing
 } from '@kepler.gl/utils';
-import {breakPointValues} from '@kepler.gl/styles';
 
 // default-settings
 import {
@@ -93,7 +92,6 @@ import {
   getStarsBackgroundImage,
   KeplerGlobeView
 } from '@kepler.gl/deckgl-layers';
-import type {GlobeAttribution} from '@kepler.gl/deckgl-layers';
 
 import {DROPPABLE_MAP_CONTAINER_TYPE} from './common/dnd-layer-items';
 // Contexts
@@ -161,23 +159,13 @@ const StyledMap = styled(StyledMapContainer)<StyledMapContainerProps>(
 const MAPBOXGL_STYLE_UPDATE = 'style.load';
 const MAPBOXGL_RENDER = 'render';
 
-type MapLibLogoProps = {
-  baseMapLibraryConfig: BaseMapLibraryConfig;
-};
-
-const MapLibLogo = ({baseMapLibraryConfig}: MapLibLogoProps) => (
-  <div className="attrition-logo">
-    Basemap by:
-    <a
-      style={{marginLeft: '5px'}}
-      className={`${baseMapLibraryConfig.mapLibCssClass}-ctrl-logo`}
-      target="_blank"
-      rel="noopener noreferrer"
-      href={baseMapLibraryConfig.mapLibUrl}
-      aria-label={`${baseMapLibraryConfig.mapLibName} logo`}
-    />
-  </div>
-);
+// Canonical OpenStreetMap attribution, injected into the collected list when a
+// style is detected to use OSM tiles but the resolved sources haven't yielded
+// an attribution string of their own (e.g. raw style declares OSM but the
+// TileJSON hasn't exposed it). Rendered through the same safe link parser as
+// every other collected attribution.
+const OSM_ATTRIBUTION_HTML =
+  '© <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a>';
 
 interface StyledDroppableProps {
   $isOver: boolean;
@@ -205,320 +193,9 @@ export const Droppable = ({containerId}) => {
   return <StyledDroppable ref={setNodeRef} $isOver={isOver} />;
 };
 
-interface StyledDatasetAttributionsContainerProps {
-  isPalm: boolean;
-}
-
-const StyledDatasetAttributionsContainer = styled.div<StyledDatasetAttributionsContainerProps>`
-  max-width: ${props => (props.isPalm ? '200px' : '300px')};
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-  color: ${props => props.theme.labelColor};
-  margin-right: 2px;
-  margin-bottom: 1px;
-  line-height: ${props => (props.isPalm ? '1em' : '1.4em')};
-
-  &:hover {
-    white-space: inherit;
-  }
-`;
-
-const DatasetAttributions = ({
-  datasetAttributions,
-  isPalm
-}: {
-  datasetAttributions: DatasetAttribution[];
-  isPalm: boolean;
-}) => (
-  <>
-    {datasetAttributions?.length ? (
-      <StyledDatasetAttributionsContainer isPalm={isPalm}>
-        {datasetAttributions.map((ds, idx) => (
-          <a
-            {...(ds.url ? {href: ds.url} : null)}
-            target="_blank"
-            rel="noopener noreferrer"
-            key={`${ds.title}_${idx}`}
-          >
-            {ds.title}
-            {idx !== datasetAttributions.length - 1 ? ', ' : null}
-          </a>
-        ))}
-      </StyledDatasetAttributionsContainer>
-    ) : null}
-  </>
-);
-
-type AttributionProps = {
-  showBaseMapLibLogo: boolean;
-  showOsmBasemapAttribution: boolean;
-  basemapAttributions?: string[];
-  datasetAttributions: DatasetAttribution[];
-  baseMapLibraryConfig: BaseMapLibraryConfig;
-  globeAttributions?: GlobeAttribution[];
-};
-
-const ATTRIBUTION_SAFE_HREF_REGEX = /^(https?:|mailto:)/i;
-
-/**
- * Strip all markup from a string as a fallback for environments without a DOM
- * (e.g. SSR). Uses a non-greedy tag matcher; only used when DOMParser is
- * unavailable, so it never has to be perfectly correct.
- */
-function stripTags(html: string): string {
-  return html.replace(/<[^>]*>/g, '').trim();
-}
-
-/**
- * Recursively convert parsed DOM nodes into safe React nodes. Only anchor
- * elements with http(s)/mailto hrefs become links; every other element is
- * flattened to its text content. Text nodes are emitted verbatim (the DOM has
- * already decoded HTML entities for us).
- */
-function domNodesToReact(
-  nodes: NodeListOf<ChildNode> | ChildNode[],
-  keyPrefix: string,
-  counter: {i: number}
-): React.ReactNode[] {
-  const result: React.ReactNode[] = [];
-  nodes.forEach(node => {
-    if (node.nodeType === 3 /* TEXT_NODE */) {
-      const text = node.textContent;
-      if (text) {
-        result.push(<React.Fragment key={`${keyPrefix}-t${counter.i++}`}>{text}</React.Fragment>);
-      }
-      return;
-    }
-    if (node.nodeType !== 1 /* ELEMENT_NODE */) {
-      return;
-    }
-    const el = node as Element;
-    if (el.tagName.toLowerCase() === 'a') {
-      const href = el.getAttribute('href') || '';
-      const text = el.textContent || '';
-      if (!text) return;
-      if (ATTRIBUTION_SAFE_HREF_REGEX.test(href)) {
-        result.push(
-          <a
-            key={`${keyPrefix}-a${counter.i++}`}
-            href={href}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            {text}
-          </a>
-        );
-      } else {
-        // drop unsafe hrefs (e.g. javascript:) but keep the link text
-        result.push(<React.Fragment key={`${keyPrefix}-a${counter.i++}`}>{text}</React.Fragment>);
-      }
-      return;
-    }
-    // any other element: keep its (recursively resolved) children
-    result.push(...domNodesToReact(el.childNodes, keyPrefix, counter));
-  });
-  return result;
-}
-
-/**
- * Parse a basemap attribution HTML string into safe React nodes. Only anchor
- * tags with http(s)/mailto hrefs become links; everything else is rendered as
- * plain text. This avoids dangerouslySetInnerHTML while preserving the links
- * custom basemap styles (e.g. OpenFreeMap) declare in their attribution.
- *
- * Uses the browser's DOMParser (inert: scripts don't run and no resources are
- * fetched) so nested tags, `>` inside attributes, unusual quoting, and HTML
- * entities are all handled correctly. Falls back to plain text where no DOM is
- * available (e.g. SSR).
- */
-export function renderBasemapAttribution(html: string, keyPrefix: string): React.ReactNode[] {
-  const DOMParserCtor =
-    (typeof window !== 'undefined' && (window as any).DOMParser) ||
-    (typeof globalThis !== 'undefined' && (globalThis as any).DOMParser) ||
-    null;
-  if (!DOMParserCtor) {
-    const text = stripTags(html);
-    return text ? [<React.Fragment key={`${keyPrefix}-t0`}>{text}</React.Fragment>] : [];
-  }
-  try {
-    const doc = new DOMParserCtor().parseFromString(html, 'text/html');
-    return domNodesToReact(doc.body.childNodes, keyPrefix, {i: 0}).filter(Boolean);
-  } catch {
-    const text = stripTags(html);
-    return text ? [<React.Fragment key={`${keyPrefix}-t0`}>{text}</React.Fragment>] : [];
-  }
-}
-
-export const Attribution: React.FC<AttributionProps> = ({
-  showBaseMapLibLogo = true,
-  showOsmBasemapAttribution = false,
-  basemapAttributions = [],
-  datasetAttributions,
-  baseMapLibraryConfig,
-  globeAttributions = []
-}: AttributionProps) => {
-  const isPalm = hasMobileWidth(breakPointValues);
-
-  const memoizedComponents = useMemo(() => {
-    // attributions declared by custom basemap styles, rendered as safe links
-    const basemapAttributionNodes = basemapAttributions.length ? (
-      <>
-        {basemapAttributions.map((attribution, idx) => (
-          <React.Fragment key={`basemap-attr-${idx}`}>
-            <span className="pipe-separator">|</span>
-            <span className="basemap-attribution">
-              {renderBasemapAttribution(attribution, `basemap-attr-${idx}`)}
-            </span>
-          </React.Fragment>
-        ))}
-      </>
-    ) : null;
-
-    // fall back to the generic OSM link only when no explicit attribution was
-    // collected from the resolved sources (avoids double-rendering)
-    const showOsmFallback = showOsmBasemapAttribution && !basemapAttributions.length;
-
-    const globeAttributionNodes = globeAttributions.length ? (
-      <>
-        {globeAttributions.map(attr => (
-          <React.Fragment key={attr.label}>
-            <span className="pipe-separator">|</span>
-            <a href={attr.href} target="_blank" rel="noopener noreferrer">
-              {attr.label}
-            </a>
-          </React.Fragment>
-        ))}
-      </>
-    ) : null;
-
-    if (!showBaseMapLibLogo) {
-      return (
-        <StyledAttribution
-          mapLibCssClass={baseMapLibraryConfig.mapLibCssClass}
-          mapLibAttributionCssClass={baseMapLibraryConfig.mapLibAttributionCssClass}
-        >
-          <EndHorizontalFlexbox>
-            <DatasetAttributions datasetAttributions={datasetAttributions} isPalm={isPalm} />
-            <div className="attrition-link">
-              {datasetAttributions?.length ? <span className="pipe-separator">|</span> : null}
-              {globeAttributions.length && isPalm ? (
-                <MapLibLogo baseMapLibraryConfig={baseMapLibraryConfig} />
-              ) : null}
-              <a href="https://kepler.gl/policy/" target="_blank" rel="noopener noreferrer">
-                © kepler.gl
-              </a>
-              {basemapAttributionNodes}
-              {globeAttributionNodes}
-              {globeAttributions.length ? <span className="pipe-separator">|</span> : null}
-              {globeAttributions.length && !isPalm ? (
-                <MapLibLogo baseMapLibraryConfig={baseMapLibraryConfig} />
-              ) : null}
-            </div>
-          </EndHorizontalFlexbox>
-        </StyledAttribution>
-      );
-    }
-
-    return (
-      <StyledAttribution
-        mapLibCssClass={baseMapLibraryConfig.mapLibCssClass}
-        mapLibAttributionCssClass={baseMapLibraryConfig.mapLibAttributionCssClass}
-      >
-        <EndHorizontalFlexbox>
-          <DatasetAttributions datasetAttributions={datasetAttributions} isPalm={isPalm} />
-          <div className="attrition-link">
-            {datasetAttributions?.length ? <span className="pipe-separator">|</span> : null}
-            {isPalm ? <MapLibLogo baseMapLibraryConfig={baseMapLibraryConfig} /> : null}
-            <a href="https://kepler.gl/policy/" target="_blank" rel="noopener noreferrer">
-              © kepler.gl
-            </a>
-            {basemapAttributionNodes}
-            {showOsmFallback ? (
-              <>
-                <span className="pipe-separator">|</span>
-                <a
-                  href="https://www.openstreetmap.org/copyright"
-                  target="_blank"
-                  rel="noopener noreferrer"
-                >
-                  © OpenStreetMap
-                </a>
-              </>
-            ) : null}
-            {globeAttributionNodes}
-            <span className="pipe-separator">|</span>
-            {!isPalm ? <MapLibLogo baseMapLibraryConfig={baseMapLibraryConfig} /> : null}
-          </div>
-        </EndHorizontalFlexbox>
-      </StyledAttribution>
-    );
-  }, [
-    showBaseMapLibLogo,
-    showOsmBasemapAttribution,
-    basemapAttributions,
-    datasetAttributions,
-    isPalm,
-    baseMapLibraryConfig,
-    globeAttributions
-  ]);
-
-  return memoizedComponents;
-};
-
-const StyledAttributionLogoContainer = styled.div<{$left: number}>`
-  position: absolute;
-  bottom: ${props => props.theme.sidePanel.margin.left}px;
-  left: ${props => props.$left}px;
-  z-index: 1;
-  display: flex;
-  align-items: flex-end;
-  gap: 4px;
-  pointer-events: auto;
-  transition: left 250ms ease-in-out;
-`;
-
-const StyledLogoLink = styled.a<{$enabled: boolean}>`
-  cursor: ${props => (props.$enabled ? 'pointer' : 'default')};
-  display: flex;
-  align-items: flex-end;
-`;
-
-type AttributionLogosProps = {
-  logos: AttributionWithStyle[];
-  activeSidePanel?: boolean;
-  sidePanelWidth?: number;
-};
-
-const LOGO_LEFT_ADJUSTMENT = 3;
-
-export const AttributionLogos: React.FC<AttributionLogosProps> = ({
-  logos,
-  activeSidePanel,
-  sidePanelWidth
-}) => {
-  const theme = useTheme() as any;
-  const left =
-    (activeSidePanel ? (sidePanelWidth || 0) + LOGO_LEFT_ADJUSTMENT : 0) +
-    theme.sidePanel.margin.left;
-
-  if (!logos?.length) return null;
-  return (
-    <StyledAttributionLogoContainer $left={left}>
-      {logos.map((logo, idx) => (
-        <StyledLogoLink
-          key={logo.logoUrl || idx}
-          href={logo.url || undefined}
-          {...(logo.url ? {target: '_blank', rel: 'noopener noreferrer'} : {})}
-          $enabled={Boolean(logo.url)}
-          style={logo.bottom ? {marginBottom: logo.bottom} : undefined}
-        >
-          <img src={logo.logoUrl} style={{height: logo.height || 12}} alt={logo.title} />
-        </StyledLogoLink>
-      ))}
-    </StyledAttributionLogoContainer>
-  );
-};
+// Attribution UI (parser + components) lives in ./map/attribution and is
+// re-exported here to preserve the public import path (@kepler.gl/components).
+export {Attribution, AttributionLogos, renderBasemapAttribution, dedupeBasemapAttributions};
 
 MapContainerFactory.deps = [MapPopoverFactory, MapControlFactory, EditorFactory];
 
@@ -613,7 +290,6 @@ export default function MapContainerFactory(
 
     state = {
       showBaseMapAttribution: true,
-      showOsmAttribution: false,
       // attribution strings declared by custom basemap styles (e.g. OpenFreeMap)
       basemapAttributions: [] as string[]
     };
@@ -643,7 +319,6 @@ export default function MapContainerFactory(
         if (this.props.mapStyle.styleType === NO_MAP_ID) {
           this.setState({
             showBaseMapAttribution: false,
-            showOsmAttribution: false,
             basemapAttributions: []
           });
         } else {
@@ -866,17 +541,21 @@ export default function MapContainerFactory(
       const usesMapbox = styleObj ? isStyleUsingMapboxTiles(styleObj) : false;
       const usesOsm = styleObj ? isStyleUsingOpenStreetMapTiles(styleObj) : false;
       const basemapAttributions = getBaseMapAttributions(this._map);
+      // if OSM tiles are detected but no source-level attribution string was
+      // collected, fold in the canonical OSM attribution so the render path
+      // stays uniform (no separate hardcoded fallback link)
+      if (usesOsm && !basemapAttributions.length) {
+        basemapAttributions.push(OSM_ATTRIBUTION_HTML);
+      }
 
       if (usesMapbox || usesOsm || basemapAttributions.length) {
         this.setState({
           showBaseMapAttribution: true,
-          showOsmAttribution: usesOsm,
           basemapAttributions
         });
       } else {
         this.setState({
           showBaseMapAttribution: false,
-          showOsmAttribution: false,
           basemapAttributions
         });
         this._checkOsmAttributionOnSourceLoad();
@@ -902,11 +581,13 @@ export default function MapContainerFactory(
         // source is loaded so custom basemap attributions are not lost.
         const basemapAttributions = getBaseMapAttributions(this._map);
         const usesOsm = mapHasOpenStreetMapAttribution(this._map);
-        if (basemapAttributions.length || usesOsm) {
+        if (usesOsm && !basemapAttributions.length) {
+          basemapAttributions.push(OSM_ATTRIBUTION_HTML);
+        }
+        if (basemapAttributions.length) {
           this._removeOsmSourceDataListener();
           this.setState({
             showBaseMapAttribution: true,
-            showOsmAttribution: usesOsm,
             basemapAttributions
           });
         } else if (attempts >= MAX_ATTEMPTS) {
@@ -1764,7 +1445,6 @@ export default function MapContainerFactory(
           {this.props.primary ? (
             <Attribution
               showBaseMapLibLogo={this.state.showBaseMapAttribution}
-              showOsmBasemapAttribution={this.state.showOsmAttribution}
               basemapAttributions={this.state.basemapAttributions}
               datasetAttributions={datasetAttributions}
               baseMapLibraryConfig={baseMapLibraryConfig}
