+++ src/utils/src/map-style-utils/mapbox-utils.ts
@@ -40,17 +40,34 @@ export function isStyleUsingOpenStreetMapTiles(mapStyle: any) {
 }
 
 /**
+ * Minimal structural contract for the MapLibre/Mapbox map instance used by the
+ * attribution helpers below. Only the members actually read are declared;
+ * everything is optional because these helpers may be called before the map is
+ * fully initialized (guarded by optional chaining + try/catch at runtime).
+ *
+ * `getSource` returns `any` on purpose: the concrete MapLibre/Mapbox source
+ * union (GeoJSONSource, VectorSource, …) is not structurally assignable to a
+ * `{attribution?}` shape, and we only ever read an optional `attribution`
+ * string off it. Declaring the two getters still catches passing a non-map.
+ */
+export interface AttributionMapLike {
+  getStyle?: () => {sources?: Record<string, unknown>} | undefined | null;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  getSource?: (id: string) => any;
+}
+
+/**
  * Checks resolved map sources for OpenStreetMap attribution.
  * CARTO basemaps don't have attribution in the raw style JSON;
  * it only appears after MapLibre resolves the TileJSON.
  * Uses map.getSource() to access the resolved source with TileJSON metadata.
  */
-export function mapHasOpenStreetMapAttribution(map: any): boolean {
+export function mapHasOpenStreetMapAttribution(map?: AttributionMapLike | null): boolean {
   try {
     const style = map?.getStyle?.();
     if (!style?.sources) return false;
     return Object.keys(style.sources).some(sourceId => {
-      const source = map.getSource?.(sourceId);
+      const source = map?.getSource?.(sourceId);
       const attribution = source?.attribution;
       if (typeof attribution === 'string') {
         return attribution.toLowerCase().includes('openstreetmap');
@@ -72,13 +89,13 @@ export function mapHasOpenStreetMapAttribution(map: any): boolean {
  * @param map the resolved MapLibre/Mapbox map instance
  * @returns de-duplicated list of attribution HTML strings, order preserved
  */
-export function getBaseMapAttributions(map: any): string[] {
+export function getBaseMapAttributions(map?: AttributionMapLike | null): string[] {
   try {
     const style = map?.getStyle?.();
     if (!style?.sources) return [];
     const seen = new Set<string>();
     Object.keys(style.sources).forEach(sourceId => {
-      const source = map.getSource?.(sourceId);
+      const source = map?.getSource?.(sourceId);
       const attribution = source?.attribution;
       if (typeof attribution === 'string') {
         const trimmed = attribution.trim();
