+++ test/browser/components/attribution-test.js
@@ -4,7 +4,7 @@
 import React from 'react';
 import test from 'tape';
 import {mountWithTheme} from 'test/helpers/component-utils';
-import {renderBasemapAttribution} from '@kepler.gl/components';
+import {renderBasemapAttribution, dedupeBasemapAttributions} from '@kepler.gl/components';
 
 // Helper: mount the parsed nodes and return the enzyme wrapper.
 function mountAttribution(html) {
@@ -19,9 +19,7 @@ test('renderBasemapAttribution -> plain text', t => {
 });
 
 test('renderBasemapAttribution -> single safe anchor', t => {
-  const wrapper = mountAttribution(
-    '© <a href="https://carto.com/attributions">CARTO</a>'
-  );
+  const wrapper = mountAttribution('© <a href="https://carto.com/attributions">CARTO</a>');
   const anchors = wrapper.find('a');
   t.equal(anchors.length, 1, 'One anchor rendered');
   t.equal(anchors.at(0).prop('href'), 'https://carto.com/attributions', 'href preserved');
@@ -81,8 +79,7 @@ test('renderBasemapAttribution -> mailto is allowed', t => {
 test('renderBasemapAttribution -> nested tags and > inside attribute (regex-breaking cases)', t => {
   // A '>' inside an attribute value and nested markup would break the old
   // regex-based parser; the DOM parser handles both.
-  const html =
-    '<a href="https://x.com" data-tip="a > b"><b>Bold</b> Link</a> tail';
+  const html = '<a href="https://x.com" data-tip="a > b"><b>Bold</b> Link</a> tail';
   const wrapper = mountAttribution(html);
   const anchors = wrapper.find('a');
   t.equal(anchors.length, 1, 'One anchor despite > inside attribute');
@@ -96,5 +93,117 @@ test('renderBasemapAttribution -> empty / whitespace', t => {
   t.deepEqual(renderBasemapAttribution('', 'k'), [], 'empty string yields no nodes');
   const wrapper = mountAttribution('   ');
   t.equal(wrapper.find('a').length, 0, 'whitespace yields no anchors');
+  t.end();
+});
+
+test('renderBasemapAttribution -> whitespace-only fragments between links are dropped', t => {
+  // whitespace-only text nodes sitting between two links should not produce
+  // empty-ish fragments
+  const html = '<a href="https://a.com">A</a>   <a href="https://b.com">B</a>';
+  const nodes = renderBasemapAttribution(html, 'k');
+  const wrapper = mountAttribution(html);
+  t.equal(wrapper.find('a').length, 2, 'both links rendered');
+  // only two nodes: the two anchors, no whitespace fragment in between
+  t.equal(nodes.length, 2, 'no whitespace-only fragment emitted between the links');
+  t.end();
+});
+
+test('renderBasemapAttribution -> nested <a> inside <a> does not produce link-in-link', t => {
+  // invalid markup: an anchor nested inside another anchor. Browsers (and
+  // JSDOM) auto-close the outer <a>, yielding sibling anchors; the parser must
+  // never emit an anchor that contains another anchor.
+  const html = '<a href="https://outer.com">outer <a href="https://inner.com">inner</a></a>';
+  const wrapper = mountAttribution(html);
+  const container = wrapper.getDOMNode();
+  const anchorEls = container.querySelectorAll('a');
+  anchorEls.forEach(a => {
+    t.equal(a.querySelectorAll('a').length, 0, 'no anchor nested inside another anchor');
+  });
+  t.ok(wrapper.text().includes('inner'), 'inner link text preserved');
+  t.ok(wrapper.text().includes('outer'), 'outer link text preserved');
+  t.end();
+});
+
+test('dedupeBasemapAttributions -> exact duplicates', t => {
+  t.deepEqual(
+    dedupeBasemapAttributions(['© OpenStreetMap', '© OpenStreetMap']),
+    ['© OpenStreetMap'],
+    'collapses exact duplicates'
+  );
+  t.end();
+});
+
+test('dedupeBasemapAttributions -> wording variants collapse (OSM vs OSM contributors)', t => {
+  t.deepEqual(
+    dedupeBasemapAttributions(['© OpenStreetMap contributors', '© OpenStreetMap']),
+    ['© OpenStreetMap contributors'],
+    'the standalone OSM string is subsumed by the fuller one'
+  );
+  t.end();
+});
+
+test('dedupeBasemapAttributions -> standalone OSM link subsumed by CARTO string that links OSM', t => {
+  const carto =
+    '© <a href="https://carto.com/attributions">CARTO</a>, © ' +
+    '<a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors';
+  const osm = '<a href="https://www.openstreetmap.org/copyright">© OpenStreetMap</a>';
+  t.deepEqual(
+    dedupeBasemapAttributions([carto, osm]),
+    [carto],
+    'OSM-only entry dropped because CARTO already links OSM and the text overlaps'
+  );
+  t.end();
+});
+
+test('dedupeBasemapAttributions -> distinct providers are all kept', t => {
+  const ofm = '<a href="https://openfreemap.org">OpenFreeMap</a>';
+  const omt = '<a href="https://www.openmaptiles.org/">© OpenMapTiles</a>';
+  const osm = '<a href="https://www.openstreetmap.org/copyright">© OpenStreetMap</a>';
+  t.deepEqual(
+    dedupeBasemapAttributions([ofm, omt, osm]),
+    [ofm, omt, osm],
+    'independent providers are preserved'
+  );
+  t.end();
+});
+
+test('dedupeBasemapAttributions -> distinct provider not dropped by token cross-contamination', t => {
+  // "Map Co" is a distinct provider; its tokens (map, co) each appear in a
+  // *different* earlier entry, but no single earlier entry covers both, so it
+  // must be kept (guards against union-based false positives)
+  const a = '© Map Data';
+  const b = '© Data Co';
+  const c = '© Map Co';
+  t.deepEqual(
+    dedupeBasemapAttributions([a, b, c]),
+    [a, b, c],
+    'entry whose tokens span two different kept entries is not subsumed'
+  );
+  t.end();
+});
+
+test('dedupeBasemapAttributions -> www prefix and casing normalized', t => {
+  const a = '<a href="https://www.OpenStreetMap.org/copyright">OpenStreetMap</a>';
+  const b = '<a href="https://openstreetmap.org/copyright">OpenStreetMap contributors</a>';
+  t.deepEqual(
+    dedupeBasemapAttributions([a, b]),
+    [a],
+    'same host regardless of www/casing collapses'
+  );
+  t.end();
+});
+
+test('dedupeBasemapAttributions -> synthesized OSM entry is subsumed by CARTO', t => {
+  // mirrors the runtime fold-in: a synthesized canonical OSM link should not
+  // double-render when a CARTO string already links OSM
+  const carto =
+    '© <a href="https://carto.com/attributions">CARTO</a>, © ' +
+    '<a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors';
+  const synthesizedOsm = '© <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a>';
+  t.deepEqual(
+    dedupeBasemapAttributions([carto, synthesizedOsm]),
+    [carto],
+    'canonical OSM entry dropped when CARTO already links OSM'
+  );
   t.end();
 });
